+++ tests/test_get_snippets.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 import sys
 from os.path import join, basename
-# Use 'io.StringIO' for Python 3.4 compatibility. In Python 2.X, still use
+# Use 'io.StringIO' for Python 3 compatibility. In Python 2, still use
 # 'StringIO.StringIO' to avoid unicode errors.
 try:
     from StringIO import StringIO