+++ update_conf_py/__init__.py
@@ -3,7 +3,7 @@
 
 from __future__ import absolute_import
 
-# Use local imports from Python 3.4 compatibility
+# Use local imports from Python 3 compatibility
 from . import main
 from .main import run